fix babel error move tremor colors under /theme Remove colors from hooks
