+++ dashboard/.eslintrc.json
@@ -1,3 +1,4 @@
 {
-  "extends": "next/core-web-vitals"
+  "extends": ["prettier", "next/core-web-vitals"],
+  "plugins": ["prettier"]
 }
